+++ array/arrOpt.cpp
@@ -1,7 +1,7 @@
 #include "arrOpt.hpp"
 #include <vector>
 
-int firstFillArray (vector<vector<int> >& arr) {
+int firstFillArray (vector<vector<long long int> >& arr) {
   int a = 0, b = 0, n = static_cast<int>(arr[0].size()), m = static_cast<int>(arr.size());
 	for (auto & i : arr) {
 		for (auto & j : i) {
@@ -11,7 +11,7 @@ int firstFillArray (vector<vector<int> >& arr) {
   return 2*n + 1;
 }
 
-int secondFillArray (vector<vector<int> >& arr) {
+int secondFillArray (vector<vector<long long int> >& arr) {
   int a = 0, b = 0, n = static_cast<int>(arr[0].size()), m = static_cast<int>(arr.size());
   for (auto & i : arr) {
     for (auto & j : i) {
